+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Deployments: purge cache endpoint
+
 ## 0.2.9
 
 * Type: Regional/apps